requirements.txt
annotated-types==0.5.0
anyio==3.7.1
beanie==1.22.6
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
ecdsa==0.18.0
fastapi==0.103.1
fastapi-jwt-auth==0.5.0
greenlet==3.0.1
h11==0.14.0
httptools==0.6.0
idna==3.4
install==1.3.5
isort==5.12.0
lazy-model==0.2.0
motor==3.3.1
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.0
pydantic==1.10.12
pydantic_core==2.10.1
PyJWT==1.7.1
pymongo==4.5.0
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==5.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
toml==0.10.2
typing_extensions==4.8.0
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
